Move validation config file to common location
